Reduce usages of IInterpreterService in web
