+++ src/notebooks/controllers/connectionDisplayData.ts
@@ -1,35 +1,16 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { injectable, inject } from 'inversify';
-import { Event, EventEmitter } from 'vscode';
-import {
-    getDisplayNameOrNameOfKernelConnection,
-    getKernelConnectionDisplayPath,
-    getKernelRegistrationInfo,
-    getRemoteKernelSessionInformation
-} from '../../kernels/helpers';
+import { EventEmitter } from 'vscode';
+import { getKernelRegistrationInfo } from '../../kernels/helpers';
 import { IJupyterServerUriStorage } from '../../kernels/jupyter/types';
 import { KernelConnectionMetadata, RemoteKernelConnectionMetadata } from '../../kernels/types';
-import { IWorkspaceService } from '../../platform/common/application/types';
-import { IPlatformService } from '../../platform/common/platform/types';
-import { IDisposable, IDisposableRegistry } from '../../platform/common/types';
+import { IDisposable } from '../../platform/common/types';
 import { DataScience } from '../../platform/common/utils/localize';
-import { noop } from '../../platform/common/utils/misc';
-import { IInterpreterService } from '../../platform/interpreter/contracts';
-import { EnvironmentType, PythonEnvironment } from '../../platform/pythonEnvironments/info';
+import { EnvironmentType } from '../../platform/pythonEnvironments/info';
+import { IConnectionDisplayData } from './types';
 
-export interface IConnectionDisplayData extends IDisposable {
-    readonly onDidChange: Event<ConnectionDisplayData>;
-    readonly connectionId: string;
-    readonly label: string;
-    readonly description: string | undefined;
-    readonly detail: string;
-    readonly category: string;
-    readonly serverDisplayName?: string;
-}
-
-class ConnectionDisplayData implements IDisposable, IConnectionDisplayData {
+export class ConnectionDisplayData implements IDisposable, IConnectionDisplayData {
     private readonly _onDidChange = new EventEmitter<ConnectionDisplayData>();
     public readonly onDidChange = this._onDidChange.event;
 
@@ -49,98 +30,8 @@ class ConnectionDisplayData implements IDisposable, IConnectionDisplayData {
     }
 }
 
-@injectable()
-export class ConnectionDisplayDataProvider {
-    private readonly details = new Map<string, ConnectionDisplayData>();
-    constructor(
-        @inject(IWorkspaceService) private readonly workspace: IWorkspaceService,
-        @inject(IPlatformService) private readonly platform: IPlatformService,
-        @inject(IJupyterServerUriStorage) private readonly serverUriStorage: IJupyterServerUriStorage,
-        @inject(IDisposableRegistry) private readonly disposables: IDisposableRegistry,
-        @inject(IInterpreterService) private readonly interpreters: IInterpreterService
-    ) {}
-
-    public getDisplayData(connection: KernelConnectionMetadata): IConnectionDisplayData {
-        if (!this.details.get(connection.id)) {
-            const label = getDisplayNameOrNameOfKernelConnection(connection);
-            const description = getKernelConnectionDisplayPath(connection, this.workspace, this.platform);
-            const detail =
-                connection.kind === 'connectToLiveRemoteKernel' ? getRemoteKernelSessionInformation(connection) : '';
-            const category = getKernelConnectionCategorySync(connection);
-            const newDetails = new ConnectionDisplayData(connection.id, label, description, detail, category);
-            this.disposables.push(newDetails);
-            this.details.set(connection.id, newDetails);
-
-            // If the interpreter information changes, then update the display data.
-            if (connection.kind === 'startUsingPythonInterpreter' && connection.interpreter.isCondaEnvWithoutPython) {
-                const updateInterpreterInfo = (e: PythonEnvironment[]) => {
-                    const changedEnv = e.find((env) => env.id === connection.interpreter?.id);
-                    const interpreter = this.interpreters.resolvedEnvironments.find((env) => env.id === changedEnv?.id);
-                    if (connection.kind === 'startUsingPythonInterpreter' && interpreter) {
-                        connection.updateInterpreter(interpreter);
-                        const newLabel = getDisplayNameOrNameOfKernelConnection(connection);
-                        const newDescription = getKernelConnectionDisplayPath(
-                            connection,
-                            this.workspace,
-                            this.platform
-                        );
-                        const newCategory = getKernelConnectionCategorySync(connection);
-                        let changed = false;
-                        if (newLabel !== newDetails.label) {
-                            newDetails.label = newLabel;
-                            changed = true;
-                        }
-                        if (newDescription !== newDetails.description) {
-                            newDetails.description = newDescription;
-                            changed = true;
-                        }
-                        if (newCategory !== newDetails.category) {
-                            newDetails.category = newCategory;
-                            changed = true;
-                        }
-                        if (changed) {
-                            newDetails.triggerChange();
-                        }
-                    }
-                };
-                this.interpreters.onDidChangeInterpreter(
-                    (e) => (e ? updateInterpreterInfo([e]) : undefined),
-                    this,
-                    this.disposables
-                );
-            }
-        }
-        const details: ConnectionDisplayData = this.details.get(connection.id)!;
-        this.details.set(connection.id, details);
-
-        if (connection.kind === 'connectToLiveRemoteKernel' || connection.kind === 'startUsingRemoteKernelSpec') {
-            getRemoteServerDisplayName(connection, this.serverUriStorage)
-                .then((displayName) => {
-                    if (details.serverDisplayName !== displayName) {
-                        details.serverDisplayName = displayName;
-
-                        details.triggerChange();
-                        return;
-                    }
-                })
-                .catch(noop);
-        }
-
-        getKernelConnectionCategory(connection, this.serverUriStorage)
-            .then((kind) => {
-                if (details.category !== kind) {
-                    details.category = kind;
-                    details.triggerChange();
-                }
-            })
-            .catch(noop);
-
-        return details;
-    }
-}
-
 // For Remote connections, check if we have a saved display name for the server.
-async function getRemoteServerDisplayName(
+export async function getRemoteServerDisplayName(
     kernelConnection: RemoteKernelConnectionMetadata,
     serverUriStorage: IJupyterServerUriStorage
 ): Promise<string> {
@@ -154,7 +45,7 @@ async function getRemoteServerDisplayName(
     return DataScience.kernelDefaultRemoteDisplayName;
 }
 
-async function getKernelConnectionCategory(
+export async function getKernelConnectionCategory(
     kernelConnection: KernelConnectionMetadata,
     serverUriStorage: IJupyterServerUriStorage
 ): Promise<string> {
