+++ src/notebooks/controllers/kernelSource/kernelSelector.unit.test.ts
@@ -35,13 +35,10 @@ import {
 import { ServiceContainer } from '../../../platform/ioc/container';
 import { EnvironmentType } from '../../../platform/pythonEnvironments/info';
 import { noop } from '../../../test/core';
-import {
-    IConnectionDisplayData,
-    ConnectionDisplayDataProvider,
-    getKernelConnectionCategorySync
-} from '../connectionDisplayData';
+import { getKernelConnectionCategorySync } from '../connectionDisplayData';
 import { CreateAndSelectItemFromQuickPick, isKernelPickItem, KernelSelector } from './kernelSelector';
 import { ConnectionQuickPickItem, IQuickPickKernelItemProvider } from './types';
+import { IConnectionDisplayData, IConnectionDisplayDataProvider } from '../types';
 
 suite('Kernel Selector', () => {
     let kernelSelector: KernelSelector;
@@ -54,7 +51,7 @@ suite('Kernel Selector', () => {
     let cancellation: CancellationTokenSource;
     let notebook: NotebookDocument;
     let workspaceService: IWorkspaceService;
-    let displayDataProvider: ConnectionDisplayDataProvider;
+    let displayDataProvider: IConnectionDisplayDataProvider;
     let pythonChecker: IPythonExtensionChecker;
     let provider: ReadWrite<IQuickPickKernelItemProvider>;
     let kernelFinder: IContributedKernelFinder<KernelConnectionMetadata>;
@@ -170,7 +167,7 @@ suite('Kernel Selector', () => {
         notebook = mock<NotebookDocument>();
         pythonChecker = mock<IPythonExtensionChecker>();
         kernelFinder = mock<IContributedKernelFinder<KernelConnectionMetadata>>();
-        displayDataProvider = mock<ConnectionDisplayDataProvider>();
+        displayDataProvider = mock<IConnectionDisplayDataProvider>();
         workspaceService = mock<IWorkspaceService>();
         onDidChangeProvider = new EventEmitter<void>();
         onDidChangeProviderStatus = new EventEmitter<void>();
@@ -219,7 +216,7 @@ suite('Kernel Selector', () => {
             recommended: undefined
         };
         when(workspaceService.isTrusted).thenReturn(true);
-        when(serviceContainer.get<ConnectionDisplayDataProvider>(ConnectionDisplayDataProvider)).thenReturn(
+        when(serviceContainer.get<IConnectionDisplayDataProvider>(IConnectionDisplayDataProvider)).thenReturn(
             instance(displayDataProvider)
         );
         when(serviceContainer.get<IPythonExtensionChecker>(IPythonExtensionChecker)).thenReturn(
